PR: Revert "[cxx] Fix return type of  start_wrapper_internal -- guint32 vs. gulong."
This reverts commit 8e02c9cf732a80b77af59b9a11fceaca7490a87c.

https://github.com/mono/mono/pull/10432#discussion_r215009888

This area of thread creation is a mess in several ways.
1. We have function pointers with almost gratuitous casts,
2. inconsistent return types (motivating function pointer casts)
3. return values ignored

https://github.com/mono/mono/issues/10467
https://github.com/mono/mono/issues/10319#issuecomment-417739621
and more.

+++ mono/metadata/threads.c
@@ -1086,8 +1086,7 @@ fire_attach_profiler_events (MonoNativeThreadId tid)
 		"Handle Stack"));
 }
 
-static gulong WINAPI
-start_wrapper_internal (StartInfo *start_info, gsize *stack_ptr)
+static guint32 WINAPI start_wrapper_internal(StartInfo *start_info, gsize *stack_ptr)
 {
 	ERROR_DECL (error);
 	MonoThreadStart start_func;
